+++ src/main/java/com/bizcub/simpleDatapacks/config/Configs.java
@@ -4,7 +4,7 @@ import com.bizcub.simpleDatapacks.SimpleDatapacks;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.ConfigData;
 import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry;
+import me.shedaniel.autoconfig.annotation.ConfigEntry.Gui.Tooltip;
 import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
 
 import java.util.ArrayList;
@@ -14,16 +14,16 @@ import java.util.List;
 @Config(name = SimpleDatapacks.modId)
 public class Configs implements ConfigData {
 
-    @ConfigEntry.Gui.Tooltip
+    @Tooltip
     public boolean copyDatapacks = false;
 
-    @ConfigEntry.Gui.Tooltip
+    @Tooltip
     public boolean showFeatures = true;
 
-    @ConfigEntry.Gui.Tooltip
+    @Tooltip
     public boolean globalDatapacks = false;
 
-    @ConfigEntry.Gui.Tooltip
+    @Tooltip
     public boolean sendRestartWarning = true;
 
     public List<String> datapacksPaths = new ArrayList<>(Arrays.asList("datapacks", System.getProperty("user.home") + "\\Downloads"));