+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* 1.21.11 update
+* Now when applying data packs, a warning appears about the need to restart the world
+* Fixed an issue where Cloth Config was required for Forge versions higher than 1.21.3